Use the correct orbital period when calculating day length.
Need to use the orbit that is a direct child of the sun to get the correct
effective day length (modulo wobble). This fixes the problem with EB
thinking sunrise is infinitely in the future for tidally locked moons
(Moon-Earth, Mun-Kerbin, etc), and probably improves EB's accuracy on moons
that are not tidally locked (Minmus-Kerbin).